[Akshay]: comments below.
The code has been revamped to create and maintain an entity list based
queue, so as to make it really asynchronous in an LB environment.

This code contains revamped backend logic and first two tabs. REmaining
tabs currently work off the static mongo data, which will start
integrating from tomorrow. So not checking in right now.