[Akshay]: QB issue 1: handle GFT or any item being added as Non Inventory Part rather than payment type.
1. Any item that is "Item Non Inventory Part" will be set to use -ve
price.
2. Gift card product code (GFT or anything as set in the settings), if
it's set as Item Non Inventory Part be used as -ve amount. Else, +ve
amount.
3. PaymentMethodRef will not be set for StoreCredit payment type. There
is a hard check in OrderHandler not to set it if payment type is
StoreCredit. So SalesReceipt in QB does not show any payment method
selected.
4. So an order paid with GFT will show total as 0.0 in QB since we are
adding GFT as -ve line item. I assume that is expected. This ordre shows
order total in POSTED tab of extension. Just wanted to point out since
order total is different in two systems.